State: Migrate billing-transactions selectors to global selectors
