+++ client/state/selectors/get-past-billing-transaction.js
@@ -1,0 +1,32 @@
+/**
+ * External dependencies
+ */
+import { get, find } from 'lodash';
+
+/**
+ * Internal dependencies
+ */
+import createSelector from 'lib/create-selector';
+import { getBillingTransactions } from './';
+
+/**
+ * Returns a past billing transaction.
+ * Returns null if the billing transactions have not been fetched yet, or there is no transaction with that ID.
+ *
+ * @param  {Object}  state   Global state tree
+ * @param  {String}  id      ID of the transaction
+ * @return {?Object}         The transaction object
+ */
+const getPastBillingTransaction = createSelector(
+	( state, id ) => {
+		const pastTransactions = get( getBillingTransactions( state ), [ 'past' ], null );
+		if ( ! pastTransactions ) {
+			return null;
+		}
+
+		return find( pastTransactions, { id } ) || null;
+	},
+	getBillingTransactions
+);
+
+export default getPastBillingTransaction;
