+++ client/state/selectors/index.js
@@ -17,15 +17,18 @@ export areSitePermalinksEditable from './are-site-permalinks-editable';
 export canCurrentUser from './can-current-user';
 export countPostLikes from './count-post-likes';
 export editedPostHasContent from './edited-post-has-content';
+export getBillingTransactions from './get-billing-transactions';
 export getMenuItemTypes from './get-menu-item-types';
 export getPostLikes from './get-post-likes';
 export getSharingButtons from './get-sharing-buttons';
 export getMediaItem from './get-media-item';
 export getMediaUrl from './get-media-url';
+export getPastBillingTransaction from './get-past-billing-transaction';
 export getSiteIconId from './get-site-icon-id';
 export getSiteIconUrl from './get-site-icon-url';
 export isAutomatedTransferActive from './is-automated-transfer-active';
 export isPrivateSite from './is-private-site';
+export isRequestingBillingTransactions from './is-requesting-billing-transactions';
 export isRequestingPostLikes from './is-requesting-post-likes';
 export isRequestingSharingButtons from './is-requesting-sharing-buttons';
 export isSavingSharingButtons from './is-saving-sharing-buttons';
